Rename async to emit
